Refactor data Folder
