+++ components/landing-page/features.tsx
@@ -3,7 +3,7 @@
 import { motion } from "framer-motion"
 import { Card, CardContent } from "@/components/ui/card"
 import Image from "next/image"
-import { features } from "@/data/features"
+import { features } from "@/data/landing-page/features"
 
 export default function Features() {
   return (
