+++ data/footer.ts
